Clarify integration test ownership
