+++ tests/pty/active_control.rs
@@ -3,7 +3,7 @@
 use proqi::{adapters::runtime::SystemIdGenerator, ports::environment::IdGenerator};
 use serde_json::Value;
 
-use super::{
+use super::support::{
     expect_command, json_command, json_input_command, raw_input_command, wait_for_control_owner,
     wait_for_path,
 };
